\name{cartilage}
\alias{cartilage}

\docType{data}

\title{
Cartilage data described in the package vignette.
}

\description{
These data are presented and analyzed in the vignette.
}

\usage{data(cartilage)}

\source{
Nissi, M. J., Mortazavi, S., Hughes, J., Morgan, P., and Ellermann, J. (2015). T2* relaxation time of acetabular and femoral cartilage with and without intra-articular Gd-DTPA2 in patients with femoroacetabular impingement. \emph{American Journal of Roentgenology}, \bold{204}(6), W695. 
}

\examples{
data(cartilage)
}

\keyword{datasets}
