% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krippendorff.R
\name{interval.dist}
\alias{interval.dist}
\title{Compute the squared difference between two scores.}
\usage{
interval.dist(x, y)
}
\arguments{
\item{x}{a score.}

\item{y}{a score.}
}
\value{
\eqn{(x-y)^2}, or 0 if \code{x} or \code{y} is \code{NA}.
}
\description{
Compute the squared difference between two scores.
}
\details{
This function computes the squared difference between two scores. This may be an appropriate distance function for the interval level of measurement. \code{NA}'s are handled gracefully.
}
\seealso{
\code{\link{nominal.dist}}, \code{\link{ratio.dist}}
}
