% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krippendorff.R
\name{nominal.dist}
\alias{nominal.dist}
\title{Apply the discrete metric to two scores.}
\usage{
nominal.dist(x, y)
}
\arguments{
\item{x}{a score.}

\item{y}{a score.}
}
\value{
0 if \code{x} is equal to \code{y} or if either is \code{NA}, 1 otherwise.
}
\description{
Apply the discrete metric to two scores.
}
\details{
This function applies the discrete metric to two scores. This may be an appropriate distance function for the nominal level of measurement. \code{NA}'s are handled gracefully.
}
\seealso{
\code{\link{interval.dist}}, \code{\link{ratio.dist}}
}
