% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cartilage}
\alias{cartilage}
\title{Data from an MRI study of hip cartilage in femoroacetabular impingement.}
\format{
A data frame having 323 rows and two columns
}
\usage{
data(cartilage)
}
\description{
This data frame has exactly two columns. The first column contains raw T2* values, the second column contrast-enhanced T2* values.
}
\references{
Nissi, M. J., Mortazavi, S., Hughes, J., Morgan, P., and Ellermann, J. (2015). T2* relaxation time of acetabular and femoral cartilage with and without intra-articular Gd-DTPA2 in patients with femoroacetabular impingement. \emph{American Journal of Roentgenology}, \bold{204}(6), W695.
}
\keyword{datasets}
